don't remove the departure time except from the last stop
In some cases, the MBTA extends a train past where the schedule says the
train should go. In that case, we get additional StopTimeUpdates past the
original last stop. Previously, the RemoveUnneededTimes filter would remove
the departure time from the scheduled last stop, even though it's not
actually the last stop. This updates the filter to only remove the departure
time if the stop is also the last in the group.