better handling of shuttles at the start of trips (#56)
Previously, the behavior for shuttles at the start of trips was to cancel the
entire trip. However, this isn't the desired behavior: trips skipped at the
beginning are generally running the second part of the trip.

The new behavior is to only skip the stop times at the beginning. If boarding
is available at the last stop in the shuttle, then that stop will not be
skipped but instead only lose its `arrival_time`.