fix: better handling of the first stop during shuttles
Previously, if a trip was being shuttled, we would always remove the arrival
time from the first StopTimeUpdate. However, if the train is past the shuttle
stops, the first stop may be a regular stop and we'd want to keep the arrival
time.

The new behavior only drops the arrival time if the directly previous stop
was a shuttled stop that didn't allow boarding.