fix: filter out empty StopTimeUpdate records
For StopTimeUpdate records in the Protobuf export, we don't include the
`status` field. This can result in records like this:

```
entity {
  id: "CR-Weekday-Fall-19-616"
  trip_update {
    trip {
      trip_id: "CR-Weekday-Fall-19-616"
      start_date: "20191210"
      route_id: "CR-Needham"
      direction_id: 1
    }
    stop_time_update {
      stop_sequence: 11
      stop_id: "South Station"
    }
  }
}

```

That StopTimeUpdate does not have any useful information, so instead we drop
it from the rendered version. This had some knock-on effects on test cases,
where we weren't providing any data for StopTimeUpdates we expected to be in
the output.