fix: don't render empty TripUpdates
Once we started filtering out some StopTimeUpdate values, it became possible
to end up with an empty list of StopTimeUpdates. Since we don't have any
updates, there's nothing to do with that TripUpdate and we drop it entirely.

This results in failures for some other tests, where there wasn't enough
information in the StopTimeUpdate to render.