fix: a couple tests after Bypass upgrade
Two failures were from a validation change to upgraded hackney: 304's
can't have a body, and our `send_resp(conn, 304, "redirect")` was
passing "redirect" as a body.

The other failure was from `Bypass.down/1` hanging. The new bypass
does some finagling with processes and sending message. The underlying
GenServer was returning {:noreply, ...} and relying on another process
to send the :ok to the calling process, but this was getting lost in
the way we were storing our Bypass expectations in an Agent. By
moving the expectations inline, Bypass.down/1 stopped hanging.