fix: require `stop_sequence` for StopTimeUpdates
Previously we merged StopTimeUpdates using a "key" of trip ID + stop
ID + stop sequence. Since stop ID and sequence are both optional (only
one needs to be present), this created the potential for duplicates:
two updates that referred to the same stop would not be merged if one
of them used a sequence and the other used a stop ID.

Given:

1. it's not universally possible to merge StopTimeUpdates without a
   stop sequence (it's required if a trip contains a loop)
2. merging updates correctly using "maybe stop ID or sequence or both"
   would require significant changes to the merging system
3. grouping and the `RemoveUnneededTimes` filter assume the presence of
   stop sequences
4. all our current input feeds either include stop sequences already or
   are planned to

...we can simplify things by explicitly requiring a stop sequence, and
only merging based on trip ID + stop sequence.