fix: potential periodic `GTFS.PickupDropOff` issue
Building the records to be inserted into ETS can take a while (over a
minute), and since all records are deleted from the table before this
work begins, the module will return `:unknown` for all lookups during
this time. This issue would occur once per hour, as per the static GTFS
refresh interval, and would result in `RemoveUnneededTimes` temporarily
not working.

This refactor applies the same approach used in `StopIDs`: build all
records up-front, then clear the table and immediately insert all the
records in one batch. This still takes a second or two, but the time
gap is much shorter, reducing the chance that an incorrect feed would
be produced and the time until a corrected feed would be produced.

Eliminating the issue entirely will require a more complex approach,
such as writing some "upsert" logic which also deletes records that
shouldn't be in the table anymore, or using Mnesia for its transaction
support.