refactor: return FeedUpdate from parsers
Previously, we returned a list of Mergeables, to pass into MergeFilter /
Table. As we look to better track latency, as well as support other
producers such as MQTT, it makes sense to wrap up the list of updates
into a top-level struct.

FeedUpdate implements the Enumerable protocol, so it works like a list
in practice. It also includes metadata for (optionally) the URL the data came from,
as well as (optionally) the timestamp that the data was received at.

Asana: https://app.asana.com/0/1205365039348894/1205365039348907/f