feat: allow configuring which Sinks accept FileTap files
FileTap is responsible for logging the files which went into
Concentrate. When we only had a single S3 sink, it was easy to send all
the FileTap files to the same sink. With the introduction of MQTT, we
don't want to send the FileTap files to the MQTT broker, so we need to
add a bit of optional configuration about which sinks accept the FileTap files.