feat(cancelled-trips): create SKIPPED StopTimeUpdates for cancelled trips (#370)
Problem:
Concentrate sometimes receives trips with a schedule_relationship of CANCELED and no StopTimeUpdates. As a result, first-party consumers such as MBTA Go and MBTA.com cannot update stop-level information for stops that a trip will skip.

Solution:
Add SKIPPED StopTimeUpdates for every stop in a trip in RTR. I figured out which stops were in a trip by writing a new ETS query in the GTFS.StopTimes GenServer