fix(stop-times): make ETS query for stop times more efficient (#371)
Problem:
After merging #370, concentrate was restarting more frequently than usual in dev. Further investigation indicated that these were out-of-memory restarts.

Solution:
The out-of-memory restarts seem to have been caused by the slowness of the new StopTimes.stops_for_trip/1 function, which queries ETS for all stop times for a given trip ID. Since the key of this table was {trip_id, stop_sequence}, the query had to match the key as part of a variable. And according to the match/2 documentation:

If the key is specified in the pattern, the match is very efficient. If the key is not specified, that is, if it is a variable or an underscore, the entire table must be searched. The search time can be substantial if the table is very large.

This PR adds a {trip_id, [{stop_sequence, ...}]} insert so looking up stop times for a trip can use the ETS key again.