fix(cancelled-trips): re-implement making skipped stop time updates for cancelled trips (#374)
Problem:
The PR that we previously deployed to add SKIPPED StopTimeUpdates for CANCELLED trips resulted in stop information "flickering" for the buses such that the stop to which a bus was headed would appear and disappear intermittently.

Solution:
This was happening because the bus_block_waiver?/2 function returned true for trips with no StopTimeUpdates, and RTR intermittently gets bus trips with no StopTimeUpdates (likely from Busloc?). Previously, these trips were most likely ignored, but now they cause us to cancel bus trips unnecessarily, which causes us to remove the stop_id for the VehiclePosition entities. This PR makes it so that we only consider bus trips with all SKIPPED StopTimeUpdates as block waivers.